split up route files and fix route reducers
