+++ shared/actions/login.js
@@ -25,7 +25,8 @@ import {chatTab, loginTab, peopleTab, isValidInitialTab} from '../constants/tabs
 import {deletePushTokenSaga} from './push'
 import {getExtendedStatus} from './config'
 import {isMobile} from '../constants/platform'
-import {appRouteTree, loginRouteTree} from '../app/routes'
+import appRouteTree from '../app/routes'
+import loginRouteTree from '../app/routes-login'
 import {pathSelector, navigateTo, navigateAppend, switchRouteDef} from './route-tree'
 import {type InitialState} from '../constants/types/config'
 import {type TypedState} from '../constants/reducer'
