+++ shared/actions/profile/index.js
@@ -12,7 +12,7 @@ import URL from 'url-parse'
 import keybaseUrl from '../../constants/urls'
 import openURL from '../../util/open-url'
 import {getPathProps} from '../../route-tree'
-import {loginRouteTree} from '../../app/routes'
+import loginRouteTree from '../../app/routes-login'
 import {navigateTo, navigateUp, switchRouteDef} from '../../actions/route-tree'
 import {loginTab, peopleTab} from '../../constants/tabs'
 import {pgpSaga} from './pgp'
