+++ shared/desktop/webpack.config.babel.js
@@ -228,7 +228,11 @@ const makeRenderThreadConfig = () => {
   // Have to inject some additional code if we're using HMR
   const HMREntries =
     flags.isHot && flags.isDev && !flags.isTreeShake
-      ? ['webpack-dev-server/client?http://localhost:4000', 'webpack/hot/only-dev-server']
+      ? [
+          'react-hot-loader/patch',
+          'webpack-dev-server/client?http://localhost:4000',
+          'webpack/hot/only-dev-server',
+        ]
       : []
 
   const makeEntries = () => {
